[rename only] {keyboard,mouse}.js
