+++ packages/driver/src/cy/commands/actions/type.js
@@ -5,7 +5,7 @@ const moment = require('moment')
 const $dom = require('../../../dom')
 const $elements = require('../../../dom/elements')
 const $selection = require('../../../dom/selection')
-const $Keyboard = require('../../../cypress/keyboard')
+const $Keyboard = require('../../../cy/keyboard')
 const $utils = require('../../../cypress/utils')
 const $actionability = require('../../actionability')
