+++ packages/driver/src/cy/mouse.js
